+++ packages/opds-web-client/src/__tests__/state-test.ts
@@ -1,14 +1,16 @@
 import { expect } from "chai";
 import { stub } from "sinon";
+import * as Redux from "redux";
+import buildInitialState, { State } from "../state";
 
-let dispatch = () => {};
-let initialState = "initial state";
-let alteredState = "state with collection and book";
+// standard redux dispatch returns the action you pass in
+let dispatch: Redux.Dispatch = action => action;
+// we have to cast the string states to type State
+let initialState = ("initial state" as unknown) as State;
+let alteredState = ("state with collection and book" as unknown) as State;
 let testState = initialState;
 import * as mergeRootProps from "../components/mergeRootProps";
 import * as store from "../store";
-
-import buildInitialState from "../state";
 
 describe("buildInitialState", () => {
   let collectionUrl = "collection url";
@@ -31,7 +33,9 @@ describe("buildInitialState", () => {
 
     storeStub = stub(store, "default").returns({
       dispatch,
-      getState: () => testState
+      getState: () => testState,
+      subscribe: stub(),
+      replaceReducer: stub()
     });
   });