+++ packages/opds-web-client/src/components/__tests__/Book-test.tsx
@@ -16,6 +16,9 @@ import DownloadButton from "../DownloadButton";
 import { mockRouterContext } from "../../__mocks__/routing";
 import { ActionsProvider } from "../context/ActionsContext";
 import buildStore from "../../store";
+import DataFetcher from "../../DataFetcher";
+import { adapter } from "../../OPDSDataAdapter";
+import ActionsCreator from "../../actions";
 
 let book: BookData = {
   id: "urn:librarysimplified.org/terms/id/3M%20ID/crrmnr9",
@@ -247,9 +250,12 @@ describe("Book", () => {
     it("shows summary, in book's language, with html stripped out and more link", () => {
       let context = mockRouterContext();
       let store = buildStore();
+      const fetcher = new DataFetcher({ adapter });
+      const actions = new ActionsCreator(fetcher);
+
       wrapper = mount(
         <Provider store={store}>
-          <ActionsProvider>
+          <ActionsProvider fetcher={fetcher} actions={actions}>
             <Book book={book} updateBook={updateBook} />
           </ActionsProvider>
         </Provider>,
@@ -277,12 +283,12 @@ describe("Book", () => {
       let epubButton = buttons.at(0);
       let mobiButton = buttons.at(1);
 
-      expect(epubButton.props().url).to.equal("secrets.epub");
-      expect(epubButton.props().mimeType).to.equal("application/epub+zip");
+      expect(epubButton.props().link.url).to.equal("secrets.epub");
+      expect(epubButton.props().link.type).to.equal("application/epub+zip");
       expect(epubButton.props().isPlainLink).to.equal(true);
 
-      expect(mobiButton.props().url).to.equal("secrets.mobi");
-      expect(mobiButton.props().mimeType).to.equal(
+      expect(mobiButton.props().link.url).to.equal("secrets.mobi");
+      expect(mobiButton.props().link.type).to.equal(
         "application/x-mobipocket-ebook"
       );
       expect(mobiButton.props().isPlainLink).to.equal(true);
@@ -328,9 +334,9 @@ describe("Book", () => {
         <Book book={bookCopy} updateBook={stub()} isSignedIn={false} />
       );
       let button = wrapper.find(DownloadButton);
-      expect(button.props().url).to.equal(link.url);
+      expect(button.props().link.url).to.equal(link.url);
       expect(button.props().title).to.equal(bookCopy.title);
-      expect(button.props().mimeType).to.equal(link.type);
+      expect(button.props().link.type).to.equal(link.type);
       expect(button.props().isPlainLink).to.equal(true);
     });