+++ packages/opds-web-client/src/components/context/ActionsContext.tsx
@@ -1,20 +1,27 @@
 import * as React from "react";
-import ActionsCreator from "../../actions";
 import useThunkDispatch from "../../hooks/useThunkDispatch";
 import DataFetcher from "../../DataFetcher";
-import { adapter } from "../../OPDSDataAdapter";
+import ActionsCreator from "../../actions";
 
 // The main context for this app's Actions.
-export const ActionsContext = React.createContext<ActionsCreator | undefined>(
+type ActionsContextType =
+  | { fetcher: DataFetcher; actions: ActionsCreator }
+  | undefined;
+export const ActionsContext = React.createContext<ActionsContextType>(
   undefined
 );
 
-export function ActionsProvider({ children }) {
-  const fetcher = new DataFetcher({ adapter });
-  const actions = new ActionsCreator(fetcher);
-
+export function ActionsProvider({
+  children,
+  actions,
+  fetcher
+}: {
+  children: React.ReactNode;
+  actions: ActionsCreator;
+  fetcher: DataFetcher;
+}) {
   return (
-    <ActionsContext.Provider value={actions}>
+    <ActionsContext.Provider value={{ actions, fetcher }}>
       {children}
     </ActionsContext.Provider>
   );
@@ -29,7 +36,8 @@ export function useActions() {
   if (typeof context === "undefined") {
     throw new Error("useActions must be used within a ActionsProvider");
   }
-  return { actions: context, dispatch };
+  const { actions, fetcher } = context;
+  return { actions, fetcher, dispatch };
 }
 
 export default ActionsContext;