+++ packages/opds-web-client/src/interfaces.ts
@@ -5,8 +5,11 @@ export type MediaType =
   | "application/kepub+zip"
   | "application/pdf"
   | "application/vnd.adobe.adept+xml"
+  | "vnd.adobe/adept+xml"
   | "application/x-mobipocket-ebook"
-  | "application/x-mobi8-ebook";
+  | "application/x-mobi8-ebook"
+  | "application/atom+xml;type=entry;profile=opds-catalog"
+  | "text/html;profile=http://librarysimplified.org/terms/profiles/streaming-media";
 
 export interface MediaLink {
   url: string;